# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Regional Carpool API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: carpool-posts
    description: Carpool board - drivers post trips
  - name: ride-requests
    description: Ride request board - passengers post requests
  - name: stats
    description: Summary statistics
  - name: medical-patients
    description: Medical transport patient registration
  - name: medical-drivers
    description: Medical transport driver registration
  - name: medical-appointments
    description: Recurring medical appointments
  - name: medical-transport
    description: Medical transport requests
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /carpool-posts:
    get:
      operationId: listCarpoolPosts
      tags: [carpool-posts]
      summary: List all carpool posts
      parameters:
        - name: fromLocation
          in: query
          schema: { type: string }
        - name: toLocation
          in: query
          schema: { type: string }
        - name: date
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of carpool posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarpoolPost"
    post:
      operationId: createCarpoolPost
      tags: [carpool-posts]
      summary: Driver posts a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarpoolPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolPost"

  /carpool-posts/{id}:
    get:
      operationId: getCarpoolPost
      tags: [carpool-posts]
      summary: Get a carpool post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Carpool post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarpoolPost"
        "404":
          description: Not found
    delete:
      operationId: deleteCarpoolPost
      tags: [carpool-posts]
      summary: Remove a carpool post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /carpool-posts/{id}/book:
    post:
      operationId: bookCarpoolPost
      tags: [carpool-posts]
      summary: Passenger books a seat ($5 fee)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: No seats available or already booked

  /ride-requests:
    get:
      operationId: listRideRequests
      tags: [ride-requests]
      summary: List all ride requests
      parameters:
        - name: fromLocation
          in: query
          schema: { type: string }
        - name: toLocation
          in: query
          schema: { type: string }
        - name: date
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of ride requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideRequest"
    post:
      operationId: createRideRequest
      tags: [ride-requests]
      summary: Passenger posts a ride request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideRequest"

  /ride-requests/{id}:
    get:
      operationId: getRideRequest
      tags: [ride-requests]
      summary: Get a ride request
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Ride request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideRequest"
        "404":
          description: Not found
    delete:
      operationId: deleteRideRequest
      tags: [ride-requests]
      summary: Remove a ride request
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /ride-requests/{id}/accept:
    post:
      operationId: acceptRideRequest
      tags: [ride-requests]
      summary: Driver accepts a ride request (10% commission)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideAcceptanceInput"
      responses:
        "201":
          description: Acceptance confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideAcceptance"
        "400":
          description: Already accepted

  /medical/patients:
    post:
      operationId: registerMedicalPatient
      tags: [medical-patients]
      summary: Register as a medical transport patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalPatientInput"
      responses:
        "201":
          description: Registration submitted for verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalPatient"
    get:
      operationId: listMedicalPatients
      tags: [medical-patients]
      summary: List all registered patients
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalPatient"

  /medical/patients/{id}:
    get:
      operationId: getMedicalPatient
      tags: [medical-patients]
      summary: Get a patient record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Patient record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalPatient"
        "404":
          description: Not found

  /medical/patients/{id}/verify:
    post:
      operationId: verifyMedicalPatient
      tags: [medical-patients]
      summary: Approve or reject a patient registration
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "200":
          description: Verification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalPatient"

  /medical/drivers:
    post:
      operationId: registerMedicalDriver
      tags: [medical-drivers]
      summary: Register as a medical transport driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalDriverInput"
      responses:
        "201":
          description: Registration submitted for verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalDriver"
    get:
      operationId: listMedicalDrivers
      tags: [medical-drivers]
      summary: List all registered drivers
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalDriver"

  /medical/drivers/{id}:
    get:
      operationId: getMedicalDriver
      tags: [medical-drivers]
      summary: Get a driver record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Driver record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalDriver"
        "404":
          description: Not found

  /medical/drivers/{id}/verify:
    post:
      operationId: verifyMedicalDriver
      tags: [medical-drivers]
      summary: Approve or reject a driver registration
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "200":
          description: Verification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalDriver"

  /medical/appointments:
    get:
      operationId: listRecurringAppointments
      tags: [medical-appointments]
      summary: List recurring appointments (optionally filter by patientId)
      parameters:
        - name: patientId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringAppointment"
    post:
      operationId: createRecurringAppointment
      tags: [medical-appointments]
      summary: Create a recurring appointment for a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringAppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringAppointment"

  /medical/appointments/{id}:
    delete:
      operationId: deleteRecurringAppointment
      tags: [medical-appointments]
      summary: Deactivate a recurring appointment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deactivated

  /medical/transport-requests:
    get:
      operationId: listMedicalTransportRequests
      tags: [medical-transport]
      summary: List all medical transport requests
      parameters:
        - name: patientId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of transport requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalTransportRequest"
    post:
      operationId: createMedicalTransportRequest
      tags: [medical-transport]
      summary: Request medical transport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalTransportRequestInput"
      responses:
        "201":
          description: Transport request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalTransportRequest"

  /medical/transport-requests/{id}:
    get:
      operationId: getMedicalTransportRequest
      tags: [medical-transport]
      summary: Get a transport request
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Transport request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalTransportRequest"
        "404":
          description: Not found

  /medical/transport-requests/{id}/assign:
    post:
      operationId: assignMedicalDriver
      tags: [medical-transport]
      summary: Assign a verified driver to a transport request
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDriverInput"
      responses:
        "200":
          description: Driver assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalTransportRequest"
        "400":
          description: Driver not verified or request already assigned

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get summary statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    CarpoolPost:
      type: object
      required: [id, driverName, driverPhone, fromLocation, toLocation, travelDate, travelTime, totalSeats, availableSeats, pricePerSeat, status, createdAt]
      properties:
        id:
          type: integer
        driverName:
          type: string
        driverPhone:
          type: string
        fromLocation:
          type: string
        toLocation:
          type: string
        travelDate:
          type: string
        travelTime:
          type: string
        totalSeats:
          type: integer
        availableSeats:
          type: integer
        pricePerSeat:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string

    CarpoolPostInput:
      type: object
      required: [driverName, driverPhone, fromLocation, toLocation, travelDate, travelTime, totalSeats]
      properties:
        driverName:
          type: string
        driverPhone:
          type: string
        fromLocation:
          type: string
        toLocation:
          type: string
        travelDate:
          type: string
        travelTime:
          type: string
        totalSeats:
          type: integer
        notes:
          type: string

    Booking:
      type: object
      required: [id, carpoolPostId, passengerName, passengerPhone, seatsBooked, totalPaid, status, createdAt]
      properties:
        id:
          type: integer
        carpoolPostId:
          type: integer
        passengerName:
          type: string
        passengerPhone:
          type: string
        seatsBooked:
          type: integer
        totalPaid:
          type: number
        status:
          type: string
        createdAt:
          type: string

    BookingInput:
      type: object
      required: [passengerName, passengerPhone, seatsBooked]
      properties:
        passengerName:
          type: string
        passengerPhone:
          type: string
        seatsBooked:
          type: integer

    RideRequest:
      type: object
      required: [id, passengerName, passengerPhone, fromLocation, toLocation, travelDate, passengerCount, offeredPrice, status, createdAt]
      properties:
        id:
          type: integer
        passengerName:
          type: string
        passengerPhone:
          type: string
        fromLocation:
          type: string
        toLocation:
          type: string
        travelDate:
          type: string
        passengerCount:
          type: integer
        offeredPrice:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
        acceptedDriverName:
          type: ["string", "null"]
        createdAt:
          type: string

    RideRequestInput:
      type: object
      required: [passengerName, passengerPhone, fromLocation, toLocation, travelDate, passengerCount, offeredPrice]
      properties:
        passengerName:
          type: string
        passengerPhone:
          type: string
        fromLocation:
          type: string
        toLocation:
          type: string
        travelDate:
          type: string
        passengerCount:
          type: integer
        offeredPrice:
          type: number
        notes:
          type: string

    RideAcceptance:
      type: object
      required: [id, rideRequestId, driverName, driverPhone, commission, netEarnings, createdAt]
      properties:
        id:
          type: integer
        rideRequestId:
          type: integer
        driverName:
          type: string
        driverPhone:
          type: string
        commission:
          type: number
        netEarnings:
          type: number
        createdAt:
          type: string

    RideAcceptanceInput:
      type: object
      required: [driverName, driverPhone]
      properties:
        driverName:
          type: string
        driverPhone:
          type: string

    Stats:
      type: object
      required: [totalCarpoolPosts, activeCarpoolPosts, totalRideRequests, activeRideRequests, totalBookings, totalAcceptances, seatsAvailable]
      properties:
        totalCarpoolPosts:
          type: integer
        activeCarpoolPosts:
          type: integer
        totalRideRequests:
          type: integer
        activeRideRequests:
          type: integer
        totalBookings:
          type: integer
        totalAcceptances:
          type: integer

    MedicalPatient:
      type: object
      required: [id, fullName, dateOfBirth, phone, address, suburb, state, postcode, medicareNumber, gpName, gpPhone, emergencyContactName, emergencyContactPhone, mobilityNeeds, verificationStatus, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        dateOfBirth:
          type: string
        phone:
          type: string
        address:
          type: string
        suburb:
          type: string
        state:
          type: string
        postcode:
          type: string
        medicareNumber:
          type: string
        gpName:
          type: string
        gpPhone:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        mobilityNeeds:
          type: string
        notes:
          type: ["string", "null"]
        verificationStatus:
          type: string
        createdAt:
          type: string

    MedicalPatientInput:
      type: object
      required: [fullName, dateOfBirth, phone, address, suburb, state, postcode, medicareNumber, gpName, gpPhone, emergencyContactName, emergencyContactPhone, mobilityNeeds]
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
        phone:
          type: string
        address:
          type: string
        suburb:
          type: string
        state:
          type: string
        postcode:
          type: string
        medicareNumber:
          type: string
        gpName:
          type: string
        gpPhone:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        mobilityNeeds:
          type: string
        notes:
          type: string

    MedicalDriver:
      type: object
      required: [id, fullName, phone, licenseNumber, vehicleType, vehicleRego, vehicleCapacity, hasWheelchairAccess, workingWithChildrenCheck, policeCheckDone, verificationStatus, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        licenseNumber:
          type: string
        vehicleType:
          type: string
        vehicleRego:
          type: string
        vehicleCapacity:
          type: integer
        hasWheelchairAccess:
          type: boolean
        workingWithChildrenCheck:
          type: boolean
        policeCheckDone:
          type: boolean
        verificationStatus:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    MedicalDriverInput:
      type: object
      required: [fullName, phone, licenseNumber, vehicleType, vehicleRego, vehicleCapacity, hasWheelchairAccess, workingWithChildrenCheck, policeCheckDone]
      properties:
        fullName:
          type: string
        phone:
          type: string
        licenseNumber:
          type: string
        vehicleType:
          type: string
        vehicleRego:
          type: string
        vehicleCapacity:
          type: integer
        hasWheelchairAccess:
          type: boolean
        workingWithChildrenCheck:
          type: boolean
        policeCheckDone:
          type: boolean
        notes:
          type: string

    RecurringAppointment:
      type: object
      required: [id, patientId, clinicName, clinicAddress, clinicSuburb, appointmentType, dayOfWeek, appointmentTime, frequencyWeeks, nextDate, active, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        clinicName:
          type: string
        clinicAddress:
          type: string
        clinicSuburb:
          type: string
        appointmentType:
          type: string
        dayOfWeek:
          type: string
        appointmentTime:
          type: string
        frequencyWeeks:
          type: integer
        nextDate:
          type: string
        active:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    RecurringAppointmentInput:
      type: object
      required: [patientId, clinicName, clinicAddress, clinicSuburb, appointmentType, dayOfWeek, appointmentTime, frequencyWeeks, nextDate]
      properties:
        patientId:
          type: integer
        clinicName:
          type: string
        clinicAddress:
          type: string
        clinicSuburb:
          type: string
        appointmentType:
          type: string
        dayOfWeek:
          type: string
        appointmentTime:
          type: string
        frequencyWeeks:
          type: integer
        nextDate:
          type: string
        notes:
          type: string

    MedicalTransportRequest:
      type: object
      required: [id, patientId, patientName, appointmentId, pickupAddress, pickupSuburb, destinationName, destinationAddress, tripDate, tripTime, status, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        appointmentId:
          type: ["integer", "null"]
        pickupAddress:
          type: string
        pickupSuburb:
          type: string
        destinationName:
          type: string
        destinationAddress:
          type: string
        tripDate:
          type: string
        tripTime:
          type: string
        returnTrip:
          type: boolean
        returnTime:
          type: ["string", "null"]
        assignedDriverId:
          type: ["integer", "null"]
        assignedDriverName:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    MedicalTransportRequestInput:
      type: object
      required: [patientId, pickupAddress, pickupSuburb, destinationName, destinationAddress, tripDate, tripTime, returnTrip]
      properties:
        patientId:
          type: integer
        appointmentId:
          type: integer
        pickupAddress:
          type: string
        pickupSuburb:
          type: string
        destinationName:
          type: string
        destinationAddress:
          type: string
        tripDate:
          type: string
        tripTime:
          type: string
        returnTrip:
          type: boolean
        returnTime:
          type: string
        notes:
          type: string

    AssignDriverInput:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: integer

    VerificationInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
        rejectionReason:
          type: string