# lib/api-zod/src/generated/types/medicalDriver.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regional Carpool API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicalDriver {
  id: number;
  fullName: string;
  phone: string;
  licenseNumber: string;
  vehicleType: string;
  vehicleRego: string;
  vehicleCapacity: number;
  hasWheelchairAccess: boolean;
  workingWithChildrenCheck: boolean;
  policeCheckDone: boolean;
  verificationStatus: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}