# lib/api-zod/src/generated/types/medicalPatient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regional Carpool API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicalPatient {
  id: number;
  fullName: string;
  dateOfBirth: string;
  phone: string;
  address: string;
  suburb: string;
  state: string;
  postcode: string;
  medicareNumber: string;
  gpName: string;
  gpPhone: string;
  emergencyContactName: string;
  emergencyContactPhone: string;
  mobilityNeeds: string;
  /** @nullable */
  notes?: string | null;
  verificationStatus: string;
  createdAt: string;
}