# lib/api-zod/src/generated/types/medicalTransportRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regional Carpool API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicalTransportRequest {
  id: number;
  patientId: number;
  patientName: string;
  /** @nullable */
  appointmentId: number | null;
  pickupAddress: string;
  pickupSuburb: string;
  destinationName: string;
  destinationAddress: string;
  tripDate: string;
  tripTime: string;
  returnTrip?: boolean;
  /** @nullable */
  returnTime?: string | null;
  /** @nullable */
  assignedDriverId?: number | null;
  /** @nullable */
  assignedDriverName?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}