# lib/api-zod/src/generated/types/recurringAppointmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Regional Carpool API
 * OpenAPI spec version: 0.1.0
 */

export interface RecurringAppointmentInput {
  patientId: number;
  clinicName: string;
  clinicAddress: string;
  clinicSuburb: string;
  appointmentType: string;
  dayOfWeek: string;
  appointmentTime: string;
  frequencyWeeks: number;
  nextDate: string;
  notes?: string;
}